Classical CNF-to-DNF matrices and checked LRAT refutation

+++ crates/logic/hol/src/row.rs
@@ -582,11 +582,11 @@ mod tests {
     use super::*;
     use covalence_lib_cbor::{Value as Cbor, from_reader, into_writer};
 
-    const fn reference(value: u64) -> Ref {
+    const fn reference(value: i32) -> Ref {
         Ref::new(value).unwrap()
     }
 
-    const fn import(value: u64) -> ImportId {
+    const fn import(value: i32) -> ImportId {
         ImportId::new(value).unwrap()
     }
